+++ src/Domain/Player/Repository/PlayerRepository.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App\Domain\Player\Repository;
+
+use ParagonIE\EasyDB\EasyDB;
+use App\Repository\Database;
+
+class PlayerRepository extends Database
+{
+    public function __construct(EasyDB $db)
+    {
+        parent::__construct($db);
+    }
+}